feat: [AB#13035] Preview Improvements, starter kits page
